bitbots_splines: adapt to code style
