+++ bitbots_splines/CMakeLists.txt
@@ -41,10 +41,10 @@ set(SOURCES
     src/Spline/Spline.cpp
     src/Utils/NewtonBinomial.cpp
     src/Utils/Combination.cpp
-    include/bitbots_splines/AbstractEngine.h
-    include/bitbots_splines/AbstractStabilizer.h
-    include/bitbots_splines/AbstractIK.h
-    include/bitbots_splines/AbstractVisualizer.h
+    include/bitbots_splines/abstract_engine.h
+    include/bitbots_splines/abstract_stabilizer.h
+    include/bitbots_splines/abstract_ik.h
+    include/bitbots_splines/abstract_visualizer.h
 )
 
 set(CODE_LIBRARIES )
